Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    long int n;
-    scanf("%ld",&n);
+    long long  n;
+    scanf("%lld",&n);
     int size=sizeof(n)*8;
     int y=n>>(size-1);
     if(n&1==1){
